Немного переделал страницу объявлений
