+++ src/App.tsx
@@ -6,7 +6,7 @@ import {
   Navigate,
 } from 'react-router-dom';
 import NavigationBar from './components/NavigationBar/NavigationBar.tsx';
-import ListingPage from './pages/ListingPage.tsx';
+import ListingPage from './pages/ListingPage/ListingPage.tsx';
 import AdvertisementDetailPage from './pages/AdvertisementDetailPage.tsx';
 // import Footer from './components/Footer/Footer.tsx';
 import OrdersPage from './pages/OrderPage/OrdersPage.tsx';
