+++ src/components/AddNewAdv/AddNewAdv.tsx
@@ -1,16 +1,24 @@
 import styles from './AddNewAdv.module.css';
 import * as React from 'react';
+import { useAppSelector } from '../../store/store.ts';
+import Skeleton from 'react-loading-skeleton';
 
 type AddNewAdvProps = {
   openModal: () => void;
 };
 
 const AddNewAdv: React.FC<AddNewAdvProps> = ({ openModal }) => {
+  const { loading } = useAppSelector((state) => state.advertisement);
+
   return (
     <div className={styles.addNew}>
-      <button className={styles.addNewAdv} onClick={openModal}>
-        Создать объявление
-      </button>
+      {loading ? (
+        <Skeleton width={196} height={39} borderRadius={20} />
+      ) : (
+        <button className={styles.addNewAdv} onClick={openModal}>
+          Создать объявление
+        </button>
+      )}
     </div>
   );
 };
