--- src/pages/ListingPage.tsx
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import Listing from '../components/Listing/Listing.tsx';
-
-const ListingPage: React.FC = () => {
-  return <Listing />;
-};
-
-export default ListingPage;
