+++ src/pages/ListingPage/ListingPage.module.css
@@ -1,0 +1,7 @@
+.listingPage {
+
+}
+
+.title {
+    text-align: center;
+}
